Springdata

+++ src/main/java/github/dwstanle/tickets/exception/ReservationNotFoundException.java
@@ -1,7 +1,7 @@
 package github.dwstanle.tickets.exception;
 
 public class ReservationNotFoundException extends RuntimeException {
-    public ReservationNotFoundException(Integer reservationId) {
+    public ReservationNotFoundException(Long reservationId) {
         super("could not find reservation '" + reservationId + "'.");
     }
 }
